src/app/components/users/users.component.ts
import { Component, OnInit } from '@angular/core';
import { User } from 'src/app/user';
import { UserService } from 'src/app/services/user.service';

@Component({
  selector: 'app-users',
  templateUrl: './users.component.html',
  styleUrls: ['./users.component.scss']
})
export class UsersComponent implements OnInit {
  users: User[] = [];

  constructor(private userService: UserService) { }

  ngOnInit(): void {
    this.getUsers();
  }

  getUsers(){
    this.userService.getUserData().subscribe(result => {
      this.users = result;
    })
  }

  addUser(user: User){
    user.index = this.users.length;
    console.log(user);
    this.users.push(user);
  }
}
